[test_nn] move the remaining pooling tests to test/nn/test_pooling
